g1_bot/cameras/camera_requirements.txt
opencv-python
numpy
protobuf
Pillow
pyrealsense2